boxing and unboxing

+++ code/src/java/pcgen/cdom/facet/analysis/TotalWeightFacet.java
@@ -64,7 +64,7 @@ public class TotalWeightFacet
 				if (eq.getChildCount() > 0)
 				{
 					totalWeight += (eq.getWeightAsDouble(pc) + eq
-							.getContainedWeight(pc).floatValue());
+							.getContainedWeight(pc));
 				}
 				else
 				{
@@ -75,12 +75,12 @@ public class TotalWeightFacet
 						// weight of 0. Feature #437410
 						firstClothing = false;
 						totalWeight += (eq.getWeightAsDouble(pc) * Math.max(eq
-								.getCarried().floatValue() - 1, 0));
+								.getCarried() - 1, 0));
 					}
 					else
 					{
 						totalWeight += (eq.getWeightAsDouble(pc) * eq
-								.getCarried().floatValue());
+								.getCarried());
 					}
 				}
 			}
